Migrated the rest of the ServerConsole commands over to ClientConsoleHost.
