+++ SS14.Server/ClientConsoleHost/Commands/ListCommands.cs
@@ -6,16 +6,16 @@ namespace SS14.Server.ClientConsoleHost.Commands
 {
     public class ListCommands : IClientCommand
     {
-        public string Command => "sv_list";
-        public string Description => "Lists all available commands.";
-        public string Help => "Outputs a list of all commands which are currently available to you, and a total command number.";
+        public string Command => "list";
+        public string Description => "Outputs a list of all commands which are currently available to you.";
+        public string Help => "list";
 
         public void Execute(IClientConsoleHost host, IPlayerSession player, string[] args)
         {
-            var builder = new StringBuilder("Available commands:\n");
+            var builder = new StringBuilder("SIDE NAME            DESC\n-------------------------\n");
             foreach (var command in host.AvailableCommands.Values)
             {
-                builder.AppendFormat("{0}: {1}\n", command.Command, command.Description);
+                builder.AppendLine($"S    {command.Command.PadRight(16)}{command.Description}");
             }
             var message = builder.ToString().Trim(' ', '\n');
             host.SendText(player, message);
