+++ SS14.Server/ClientConsoleHost/Commands/SpawnCommand.cs
@@ -1,4 +1,4 @@
-using SS14.Server.Interfaces.ClientConsoleHost;
+﻿using SS14.Server.Interfaces.ClientConsoleHost;
 using SS14.Server.Interfaces.GameObjects;
 using SS14.Server.Interfaces.Player;
 using SS14.Shared.IoC;
@@ -9,7 +9,7 @@ namespace SS14.Server.ClientConsoleHost.Commands
     {
         public string Command => "spawn";
         public string Description => "Spawns an entity with specific type at your feet.";
-        public string Help => "Usage: spawn <entity type>";
+        public string Help => "spawn <entity type>";
 
         public void Execute(IClientConsoleHost host, IPlayerSession player, string[] args)
         {
