+++ SS14.Server/SS14.Server.csproj
@@ -137,8 +137,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reflection\ServerReflectionManager.cs" />
     <Compile Include="BaseServer.cs" />
-    <Compile Include="ServerConsole\Commands\LogCommands.cs" />
-    <Compile Include="ServerConsole\Commands\MapCommands.cs" />
+    <Compile Include="ClientConsoleHost\Commands\LogCommands.cs" />
     <Compile Include="Signals.cs" />
     <Content Include="server_config.toml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
@@ -149,7 +148,6 @@
     <Compile Include="Chat\ChatManager.cs" />
     <Compile Include="ClientConsoleHost\Commands\ListCommands.cs" />
     <Compile Include="ClientConsoleHost\ClientConsoleHost.cs" />
-    <Compile Include="ClientConsoleHost\Commands\DrugsCommand.cs" />
     <Compile Include="GameObjects\ServerComponentFactory.cs" />
     <Compile Include="GameObjects\ServerEntityManager.cs" />
     <Compile Include="GameObjects\ServerEntityNetworkManager.cs" />
@@ -188,13 +186,6 @@
     <Compile Include="Player\PlayerManager.cs" />
     <Compile Include="Player\PlayerSession.cs" />
     <Compile Include="ServerConsole\ConsoleManager.cs" />
-    <Compile Include="ServerConsole\Commands\HelpCommand.cs" />
-    <Compile Include="ServerConsole\Commands\ListCommands.cs" />
-    <Compile Include="ServerConsole\Commands\ListPlayers.cs" />
-    <Compile Include="ServerConsole\Commands\SaveConfig.cs" />
-    <Compile Include="ServerConsole\Commands\ServerManagement.cs" />
-    <Compile Include="ServerConsole\Commands\TestCommand.cs" />
-    <Compile Include="ServerConsole\Commands\ThrowException.cs" />
     <Compile Include="Player\PlayerInput.cs" />
     <Compile Include="Interfaces\Player\IPlayerInput.cs" />
   </ItemGroup>
