--- SS14.Server/ServerConsole/Commands/MapCommands.cs
@@ -1,106 +1,0 @@
-﻿using SS14.Server.Interfaces.Maps;
-using SS14.Server.Interfaces.Player;
-using SS14.Server.Interfaces.ServerConsole;
-using SS14.Shared.Interfaces.Map;
-using SS14.Shared.IoC;
-using SS14.Shared.Map;
-
-namespace SS14.Server.ServerConsole.Commands
-{
-    public class SaveBp : IConsoleCommand
-    {
-        public string Command => "savebp";
-        public string Description => "Serializes a grid to disk.";
-        public string Help => "savebp <gridID> <Path>";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            if (args.Length < 1)
-                return;
-
-            if (!int.TryParse(args[0], out var intGridId))
-                return;
-
-            var gridId = new GridId(intGridId);
-
-            var mapManager = IoCManager.Resolve<IMapManager>();
-
-            // no saving default grid
-            if (!mapManager.TryGetGrid(gridId, out var grid) || grid.IsDefaultGrid)
-                return;
-
-            // TODO: Parse path
-            IoCManager.Resolve<IMapLoader>().SaveBlueprint(gridId, "Maps/Demo/DemoGrid.yaml");
-        }
-    }
-
-    public class LoadBp : IConsoleCommand
-    {
-        public string Command => "loadbp";
-        public string Description => "Loads a blueprint from disk into the game.";
-        public string Help => "loadbp <MapID> <GridID> <Path>";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            //TODO: Make me work after placement can create new grids.
-        }
-    }
-
-    public class SaveMap : IConsoleCommand
-    {
-        public string Command => "savemap";
-        public string Description => "Serializes a map to disk.";
-        public string Help => "savemap <MapID> <Path>";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            if (args.Length < 1)
-                return;
-
-            if (!int.TryParse(args[0], out var intMapId))
-                return;
-
-            var mapID = new MapId(intMapId);
-
-            // no saving null space
-            if (mapID == MapId.Nullspace)
-                return;
-
-            var mapManager = IoCManager.Resolve<IMapManager>();
-            if (!mapManager.TryGetMap(mapID, out var map))
-                return;
-
-            // TODO: Parse path
-            IoCManager.Resolve<IMapLoader>().SaveMap(map, "Maps/Demo/DemoMap.yaml");
-        }
-    }
-
-    public class LoadMap : IConsoleCommand
-    {
-        public string Command => "loadmap";
-        public string Description => "Loads a map from disk into the game.";
-        public string Help => "loadmap <MapID> <Path>";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            if (args.Length < 1)
-                return;
-
-            if (!int.TryParse(args[0], out var intMapId))
-                return;
-
-            var mapID = new MapId(intMapId);
-
-            // no loading null space
-            if (mapID == MapId.Nullspace)
-                return;
-
-            var mapManager = IoCManager.Resolve<IMapManager>();
-            if (mapManager.MapExists(mapID))
-                return;
-
-            // TODO: Parse path
-            IoCManager.Resolve<IMapLoader>().LoadMap(mapID, "Maps/Demo/DemoMap.yaml");
-        }
-    }
-}
