--- SS14.Server/ServerConsole/Commands/SaveConfig.cs
@@ -1,19 +1,0 @@
-﻿using SS14.Server.Interfaces.Player;
-using SS14.Shared.Interfaces.Configuration;
-using SS14.Server.Interfaces.ServerConsole;
-using SS14.Shared.IoC;
-
-namespace SS14.Server.ServerConsole.Commands
-{
-    public class SaveConfig : IConsoleCommand
-    {
-        public string Command => "saveconfig";
-        public string Description => "Saves the server configuration to the config file";
-        public string Help => "No arguments required. Saves the server configuration to the config file.";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            IoCManager.Resolve<IConfigurationManager>().SaveToFile();
-        }
-    }
-}
