--- SS14.Server/ServerConsole/Commands/ServerManagement.cs
@@ -1,32 +1,0 @@
-﻿using SS14.Server.Interfaces;
-using SS14.Server.Interfaces.Player;
-using SS14.Server.Interfaces.ServerConsole;
-using SS14.Shared.IoC;
-
-namespace SS14.Server.ServerConsole.Commands
-{
-    public class RestartServer : IConsoleCommand
-    {
-        public string Command => "restart";
-        public string Description => "Restarts the server";
-        public string Help => "Restarts the server.";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            IoCManager.Resolve<IBaseServer>().Restart();
-        }
-    }
-
-    // Crashes for some reason.
-    public class StopServer : IConsoleCommand
-    {
-        public string Command => "shutdown";
-        public string Description => "Stops the server";
-        public string Help => "Stops the server brutally without telling clients.";
-
-        public void Execute(IConsoleManager host, IPlayerSession player, string[] args)
-        {
-            IoCManager.Resolve<IBaseServer>().Shutdown();
-        }
-    }
-}
